chore: add middlewares for express, koa, hapi
