--- lerna.json
@@ -1,7 +1,0 @@
-{
-  "lerna": "2.0.0-rc.5",
-  "packages": [
-    "middlewares/graphql-voyager-*"
-  ],
-  "version": "0.0.0"
-}
