--- middlewares/graphql-voyager-express-middleware/package.json
@@ -1,31 +1,0 @@
-{
-  "name": "graphql-voyager-express-middleware",
-  "version": "0.0.1",
-  "description": "",
-  "main": "lib/index.js",
-  "scripts": {
-    "precompile": "rm -rf lib",
-    "compile": "tsc",
-    "prepublish": "npm run compile"
-  },
-  "author": "S Teplov <teplov.s.dp@gmail.com>",
-  "license": "MIT",
-  "typings": "lib/index.d.ts",
-  "typescript": {
-    "definition": "dist/index.d.ts"
-  },
-  "dependencies": {
-    "graphql-voyager": "^1.0.0-rc.3"
-  },
-  "devDependencies": {
-    "@types/express": "^4.0.36",
-    "@types/node": "^8.0.10",
-    "babel-cli": "^6.24.1",
-    "express": "^4.15.3",
-    "express-graphql": "^0.6.6",
-    "graphql": "^0.10.3"
-  },
-  "optionalDependencies": {
-    "@types/express": "^4.0.36"
-  }
-}
