--- middlewares/graphql-voyager-express-middleware/tsconfig.json
@@ -1,17 +1,0 @@
-{
-  "extends": "../../tsconfig",
-  "compilerOptions": {
-    "outDir": "./lib",
-    "types": [
-      "@types/node"
-    ]
-  },
-  "include": [
-    "../../manual-types.d.ts",
-    "./src/**/*"
-  ],
-  "exclude": [
-    "node_modules",
-    "lib"
-  ]
-}
