--- middlewares/graphql-voyager-hapi-middleware/.npmignore
@@ -1,6 +1,0 @@
-node_modules
-.tmp
-lib
-
-stats.json
-npm-debug.log
