--- middlewares/graphql-voyager-hapi-middleware/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "graphql-voyager-hapi-middleware",
-  "version": "0.0.1",
-  "description": "",
-  "main": "lib/index.js",
-  "scripts": {
-    "precompile": "rm -rf lib",
-    "compile": "tsc",
-    "prepublish": "npm run compile"
-  },
-  "author": "S Teplov <teplov.s.dp@gmail.com>",
-  "license": "MIT",
-  "typings": "lib/index.d.ts",
-  "typescript": {
-    "definition": "lib/index.d.ts"
-  },
-  "dependencies": {
-    "graphql-voyager": "^1.0.0-rc.3",
-    "hapi": "^16.4.3",
-    "hapi-graphql": "^1.0.1",
-    "inert": "^4.2.0"
-  },
-  "devDependencies": {
-    "@types/node": "^8.0.10",
-    "graphql": "^0.10.3",
-    "graphql-server-hapi": "^1.0.0",
-    "hapi": "^16.4.3"
-  },
-  "optionalDependencies": {
-    "@types/hapi": "^16.1.4"
-  }
-}
