--- middlewares/graphql-voyager-hapi-middleware/tsconfig.json
@@ -1,19 +1,0 @@
-{
-  "extends": "../../tsconfig",
-  "compilerOptions": {
-    "outDir": "./lib",
-    "types": [
-      "@types/node"
-    ]
-  },
-  "include": [
-    "../../manual-types.d.ts",
-    "./src/**/*",
-    "./example/**/*",
-    "./package.json"
-  ],
-  "exclude": [
-    "node_modules",
-    "lib"
-  ]
-}
