--- middlewares/graphql-voyager-koa-middleware/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "graphql-voyager-koa-middleware",
-  "version": "0.0.1",
-  "description": "",
-  "main": "lib/index.js",
-  "scripts": {
-    "precompile": "rm -rf lib",
-    "compile": "tsc",
-    "prepublish": "npm run compile"
-  },
-  "author": "S Teplov <teplov.s.dp@gmail.com>",
-  "license": "MIT",
-  "typings": "lib/index.d.ts",
-  "typescript": {
-    "definition": "lib/index.d.ts"
-  },
-  "dependencies": {
-    "graphql-voyager": "^1.0.0-rc.3",
-    "koa-router": "^7.2.1",
-    "koa-send": "^4.1.0"
-  },
-  "devDependencies": {
-    "@types/koa": "^2.0.39",
-    "@types/node": "^8.0.10",
-    "graphql": "^0.10.3",
-    "koa": "^2.3.0",
-    "koa-graphql": "^0.7.1"
-  },
-  "optionalDependencies": {
-    "@types/koa": "^2.0.39"
-  }
-}
