--- middlewares/graphql-voyager-koa-middleware/tsconfig.json
@@ -1,18 +1,0 @@
-{
-  "extends": "../../tsconfig",
-  "compilerOptions": {
-    "target": "es2017",
-    "outDir": "./lib",
-    "types": [
-      "@types/node"
-    ]
-  },
-  "include": [
-    "../../manual-types.d.ts",
-    "./src/**/*"
-  ],
-  "exclude": [
-    "node_modules",
-    "lib"
-  ]
-}
