+++ package.json
@@ -8,8 +8,10 @@
     "start": "webpack-dev-server --config build/webpack.demo.js",
     "clean:release": "rimraf dist",
     "clean:demo": "rimraf demo-dist",
+    "clean:middleware": "rimraf middleware",
     "build:release": "npm run clean:release && NODE_ENV=production webpack --config build/webpack.js && rm dist/voyager.worker.js.map",
     "build:demo": "npm run clean:demo && NODE_ENV=production webpack -p --config build/webpack.demo.js",
+    "build:middleware": "npm run clean:middleware && tsc src/middleware/index.ts --outDir middleware --lib ES6,DOM",
     "deploy": "deploy-to-gh-pages --local --update demo-dist",
     "stats": "NODE_ENV=production webpack --json -p --config build/webpack.js > stats.json",
     "prepare-ghpages": "node build/prepare-ghpages.js"
@@ -30,12 +32,14 @@
     "css-loader": "^0.28.4",
     "deploy-to-gh-pages": "^1.3.6",
     "ejs-compiled-loader": "2.2.0",
+    "express": "^4.15.3",
     "extract-text-webpack-plugin": "^2.1.2",
     "graphql": "^0.10.3",
+    "hapi": "^16.4.3",
     "hint.css": "^2.5.0",
     "html-webpack-plugin": "^2.29.0",
     "json-loader": "^0.5.4",
-    "lerna": "^2.0.0-rc.5",
+    "koa": "^2.3.0",
     "postcss-cssnext": "^3.0.0",
     "postcss-import": "^10.0.0",
     "postcss-loader": "^2.0.6",
