+++ src/middleware/index.ts
@@ -1,0 +1,9 @@
+import { default as express } from './express';
+import { default as hapi } from './hapi';
+import { default as koa } from './koa';
+
+export {
+  express,
+  hapi,
+  koa
+}
